type

+++ src/model.py
@@ -273,7 +273,7 @@ class SummarizationModel(object):
         emb_dec_inputs = [tf.nn.embedding_lookup(embedding, x) for x in tf.unstack(self._dec_batch, axis=1)] # list length max_dec_steps containing shape (batch_size, emb_size)
 
       # Add the encoder.
-      enc_outputs, fw_st, bw_st = self._add_encoder(emb_enc_inputs, self._enc_lens, self._enc_batch, self._enc_padding_mask)
+      enc_outputs, fw_st, bw_st = self._add_encoder(emb_enc_inputs, self._enc_lens, self._enc_batch, tf.cast(self._enc_padding_mask,tf.int32))
       self._enc_states = enc_outputs
 
       # Our encoder is bidirectional and our decoder is unidirectional so we need to reduce the final encoder hidden state to the right size to be the initial decoder hidden state
